GUACAMOLE-249: Refactor away old stream.h and guac_rdp_stream.
